KDateComboBox: Ignore invalid dates entered in date picker
This fixes the problem that potentially invalid dates emitted by
the date picker with the KDatePicker::dateEntered() signal were taken
over by the date combobox without validation. Using the same signal
handler (renamed to datePicked) as for the KDatePicker::tableClicked()
signal fixes this.

This commit also adds a small test app for KDateComboBox.